requirements.txt
# Azure Document Intelligence
azure-ai-formrecognizer==3.3.0

# Google Cloud Document AI
google-cloud-documentai==2.22.0

# Data processing and analysis
pandas==2.1.4
numpy==1.25.2

# Environment management
python-dotenv==1.0.0

# JSON and file handling (built-in modules, but listing for clarity)
# json - built-in
# os - built-in
# time - built-in
# typing - built-in
# datetime - built-in

# Optional: For additional data visualization
matplotlib==3.8.2
seaborn==0.13.0

# Optional: For more advanced document processing
pillow==10.1.0
PyPDF2==3.0.1